==> script_old/00_setup.R <==
library(fs)
library(readr)
library(dplyr)
library(lubridate)
library(tidyr)
library(stringr)
library(tidyverse)
library(dbscan)
library(purrr)
library(grid)
library(cowplot)
library(ggpubr)
library(scales)
library(highcharter)
library(gt)
library(dunn.test)
library(lsmeans)
library(rcompanion)
library(readxl)
library(Metrics)
library(sf)
library(terra)
options(timeout = max(3600, getOption("timeout")))
modal <- function(x) {
  tabla_frecuencia <- table(x)
  moda <- as.numeric(names(tabla_frecuencia)[which.max(tabla_frecuencia)])
  return(moda)
}
pca <- function(x1,x2,x3) {
  df <- data.frame(x1,x2,x3)
  pca_result <- predict(prcomp(df,scale. = T))
  return(data.frame(pca_result))
}
